pyproject.toml
[tool.poetry]
name = "lalan-midtrans-api"
version = "0.1.0"
description = "API for processing payments and retrieving transaction statuses in the Lalan app"
authors = ["Braiyen Massora <braiyenmassora@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pendulum = "^3.0.0"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
motor = "^3.5.1"
pydantic = "^2.8.2"
ngrok = "^1.3.0"
midtransclient = "^1.4.2"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
